requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
python-multipart>=0.0.5,<0.1.0
jinja2>=3.0.1,<4.0.0
pandas>=1.3.0,<2.0.0
numpy>=1.21.0,<2.0.0
transformers>=4.30.0,<5.0.0
torch>=2.0.0
python-dotenv>=0.19.0,<0.20.0
requests>=2.26.0,<3.0.0
aiofiles>=0.7.0,<0.8.0
scikit-learn>=1.2.0
bert-score>=0.3.13
shap>=0.41.0
lime>=0.2.0.1
plotly>=5.15.0
dash>=2.10.0
dash-bootstrap-components>=1.4.0 